requirements.txt
pytesseract==0.3.10
pdf2image==1.17.0
PyMuPDF==1.24.9
Pillow==10.4.0
numpy==1.26.4
opencv-python-headless==4.10.0.84
mutagen==1.47.0

